exercise: digit multiplication

+++ Behavioral/Exersice/src/main/java/org/example/Main.java
@@ -7,16 +7,16 @@ public class Main {
     public static void main(String[] args) {
         //Digit multiplication
         Scanner scanner = new Scanner(System.in);
-        int n = Integer.parseInt(scanner.next());
+        Long n = Long.parseLong(scanner.next());
         while (n > 9){
             n = divideToOneDigit(n);
         }
         System.out.println(n);
     }
-    public static int divideToOneDigit(int a){
-        int result = a % 10;
+    public static Long divideToOneDigit(Long a){
+        Long result = a % 10;
         if(result == 0){
-            return 0;
+            return 0L;
         }
         a = a / 10;
         while (a % 10 > 0){
